Fixes #2
Create the User Dashboard Layout

Implemented the basic layout for the User's dashboard. Created a standard dashboard layout, consisting of
the navigation in the Sidebar and the main layout next to the sidebar.

In the user's projects list layout, added two different views (list and card) for the user to view information
related to all their projects. User can switch between the list and the card views seamlessly. Added a list of
actions that the user can perform on the project, for example,
1. Delete the project
2. Open the project settings modal and edit the project details