fix: Added new context to manage user data
Before the changes, an authenticated user's data was being fetched
directly in the Layout component. This was fine, when only the user's
name was being used to get and display their initials.

Now, a user can be a part of many organizations, multiple projects
across Organizations, a new `UserProvider` context has been added
to be single source of truth for user profile info for the whole
application.
Also changing `UserProfile` type to `TAuthenticatedUser` to reflect new
user profile data.