feat: Implement User Invitation Flow
Each user that registers on the platform, has an organization created
for them. They are the owners of the organization.

Now, a user can invite other users to their organization. The invited
user can be someone who -
1. Is already registered on the platofrm.
2. Is not registered on the platform.

> Invited User is not registered on the platform.

**What are the steps to complete the invitation ?**
1. User sends an invitation email to the new user.
2. New user clicks on the link provided in the email
3. They are shown a joining form, where they need to enter
their `name`, `Organization name` and their `password`.
4. Upon entering their details, their account is created.

**What happens after invited user submits their details?**
1. User profile is created. No account verification email is sent
2. New `Organization` is created for the user with admin access
3. User is added to the invitation Organization's members list.

> Invited User is registered on the platform

If the invited user is already has an account, they are simply added
to the invited Organization's members list, with their role being
`MEMBER`