Add default return type for useClient hook
When using the `useClient` hook, we specify what return type is
to be expected, like `useClient<TUser>`. This ensures that we know
the response type wherever this hook is used.

Simple string responsed from the backend are transformed to this type
`{ message: Message<string> }`. So, it becomes redundant when we have
to specify the return type for these API's manually everywhere in the
frontend.

To tackle this, I have added a custom new type `Default`, which simply
checks if the type specified in `useClient` is `unknown` or not.
If it is unknown, then the return type of the hook is by default made to
be `{ message: string }`.