Added new admin action in User Admin
Added a new admin action, `send_verification_email` to the user admin.
Using this action, admin users can resend account activation emails to
users.

The admin action can only be performed on 1 user and the user should
not be verified, otherwise an error will be shown and no email will be
sent