Add task details in TaskAside in a structured way
1. Changed `currentSelectedTask` to `selectedTaskId`. Previously,
when a user was selecting a particular task, the whole task object
was being stored in state.
This means, we are storing redundant data as the selected task data
is also stored in the `tasks` array of a project.
Now, we are storing the `id` of the selected task and getting the task
data by filtering out the selected task from the `tasks` array.

2. In the task sidebar, now we are showing some of the task information
as a table. User will be able to update the data based on the type of
data they are seeing.

3. Added a new component `Dropdown` that takes in a list of values
and shows them as a dropdown. User can send an `onOptionSelect`
callback to decide what happens when a new value is selected.

4. The `priority` of a task can only be 1 value. So, it is shown as a
dropdown in the task aside sidebar. If user changes the priority,
it will be updated in the backend and success message is shown.