Show a project's assigned users in the frontend
In the single project page, we are now showing all the
users who have been assigned to the  project.

Created a new component `UserAvatar` to show the
user's profile picture if there is one or show a generic
facebook avatar.

Using it in the profile page to show the user's profile
picture.