Modified `TaskAside`
Changed the layout of the component to a 2 column
layout.

1. On the left hand side, the task description is shown
which the user can change.

2. On the right side, some of the tasks info are being shown.
For date fields, date input is shown and the user can change the date
using the native browser date picker.

Created a function `generateTaskMetaComponents` which returns an object
of the type `{ label: string; Element: JSX.Element; }`.
There are multiple types of values for task fields, such as "date",
"text", "dropdown", etc.  to show the appropriate components for each
field type, this method uses `switch` statements to generate components
for the different types of fields of a task.