Fixes #18
**Issue**
While changing the end date for a `Task`, django validates
the end date and checks if the value is after the end date or not.
If `end_date` < `task.start_date`, then an error is raised.
But in the event that the start date is not set, then this error is
raised.

**Solution**
In the validation method `validate_end_date` of the
`TasksWriteUpdateSerializer` serializer, added a condition to the
existing condition to also check if `task.start_date` has any value or
not.
If `task.start_date` has some value, only then check the end date
condition, otherwise return the value of `end_date`.

**Other Changes**
1. Added server side validation for `start_date` as well to check if the
value is before the `end_date` or not. If `start_date` > `end_date`,
then raise an error.
2. Changed settings for react toastify to show new notifications on top
of the old ones.