Moved NewTaskModal component to the page component
Previously,
the opening and closing of the NewTaskModal
component was triggered from the `kanban-column`
component. Because of this, the modal could not opened
when the tasks were shown in list mode.

Currently,
Moved the NewTaskModal component from `kanban-column`
component to the `single-project` page component and
also moved the `showNewTaskForm` state there as well.

Now, the values `showNewTaskForm` and `setShowNewTaskForm`
can be passed from the `singleProject` component to both the
`TasksKanbanView` and `TasksListView` components.