New Sidebar to allow users to export project tasks
Added a new option for the users to export their project tasks in csv
or excel format in the frontend.

To export project tasks, select the export tasks setting for the
project, a sidebar is shown with all the task fields that can be
exported. User can select all the fields they want to export
in the csv/excel file.

As for the sidebar, the width of the sidebar has been reduced.